feat: moved reflection specific code to new ReflectionUtils class
